Updates from PR comments

+++ tools/Azure.Mcp.Tools.EventHubs/tests/Azure.Mcp.Tools.EventHubs.UnitTests/Namespace/NamespaceGetCommandTests.cs
@@ -105,12 +105,12 @@ public class NamespaceGetCommandTests
         // Assert
         if (shouldSucceed)
         {
-            Assert.Equal(200, response.Status);
+            Assert.Equal(200, (int)response.Status);
             Assert.NotNull(response.Results);
         }
         else
         {
-            Assert.NotEqual(200, response.Status);
+            Assert.NotEqual(200, (int)response.Status);
             Assert.NotNull(response.Message);
         }
     }
@@ -132,7 +132,7 @@ public class NamespaceGetCommandTests
         var response = await _command.ExecuteAsync(_context, parseResult);
 
         // Assert
-        Assert.NotEqual(200, response.Status);
+        Assert.NotEqual(200, (int)response.Status);
         Assert.NotNull(response.Message);
     }
 
@@ -152,7 +152,7 @@ public class NamespaceGetCommandTests
         var response = await _command.ExecuteAsync(_context, parseResult);
 
         // Assert
-        Assert.NotEqual(200, response.Status);
+        Assert.NotEqual(200, (int)response.Status);
         Assert.NotNull(response.Message);
     }
 
@@ -168,7 +168,7 @@ public class NamespaceGetCommandTests
         var response = await _command.ExecuteAsync(_context, parseResult);
 
         // Assert
-        Assert.Equal(200, response.Status);
+        Assert.Equal(200, (int)response.Status);
         Assert.Null(response.Results); // Should be null when no namespaces found
 
         await _eventHubsService.Received(1).GetNamespacesAsync("rg-empty", "12345678-1234-1234-1234-123456789012", Arg.Any<string?>(), Arg.Any<RetryPolicyOptions?>());
@@ -199,7 +199,7 @@ public class NamespaceGetCommandTests
         var response = await _command.ExecuteAsync(_context, parseResult);
 
         // Assert
-        Assert.Equal(200, response.Status);
+        Assert.Equal(200, (int)response.Status);
         Assert.NotNull(response.Results);
 
         // Verify response structure
@@ -252,7 +252,7 @@ public class NamespaceGetCommandTests
         var response = await _command.ExecuteAsync(_context, parseResult);
 
         // Assert
-        Assert.Equal(200, response.Status);
+        Assert.Equal(200, (int)response.Status);
         Assert.NotNull(response.Results);
 
         // Since we mocked the service to return the expected namespace,
